Agents, actions, path
Created classes for agents including Students and Admin and the Player along with paths for the AI Students and Admin. Created functionality to handle directional inputs on keyboard with arrows or WASD for the player. Also included functionality for doing a left click and activating any button that the mouse location collides with however the current button list is empty.